refactor: move and rename accounts and proposals utils to services
